Split both bufferEntries tests to different files
I don't know why the test scheduler is breaking when these two tests are in the same file.
